Fixed type in findById

+++ app/src/main/java/com/muflone/android/django_hotels/database/dao/BrandDao.java
@@ -17,7 +17,7 @@ public interface BrandDao {
     List<Brand> getAll();
 
     @Query("SELECT * FROM brands WHERE id = :id")
-    Brand findById(int id);
+    Brand findById(long id);
 
     @Query("SELECT * FROM brands WHERE name = :name")
     Brand findByName(String name);
